requirements.txt
PyQt6>=6.4.0
Pillow>=9.3.0
imagehash>=4.3.1
mss>=7.0.1
python-dotenv>=0.21.0
openai>=1.3.0
elevenlabs>=0.2.24
